fix: RTD, unicode and PDF problem
Previous commit cannot be built by RTD because as of today they
hardcode the use of pdflatex. It worked fine locally. But again
locally eveything works like a charm.

Best solution seems to use \\'e and to revert to our previosu state.
I'll do it in several commit to see if forcing inputenc and utf8extra
to unset values was really necessary.