fix: one small unicode typo
Last commit checked that latex_documents was the right place to update the name of the latex document for ReadTheDocs's latex document. Note that previous commit had a broken PDF for quite some time on RTD which might be caused by some cache somewhere (or an error on my side). Anyway, current PDF version is working fine both locally and on RTD, so it's time to correct the small typo.