fix: no kwargs + avoid default parameter in LabelledChemical
We do not want kwargs in LabelledChemical because it could hide
typos in parameters names and fall back to a default value without
any warning/error.

* removed the kwargs in LabelledChemical
  * consequently, TypeError if parameter is unknown
  * implemented the associated test
* modified the CorrectorFactory so it give only the right parameters
* implemented a check in the Factory to fail on unused args
  * implemented associated test

Moreover, we do not want default parameters in LabelledChemical
because "explicit is better than implicit". Also there is a risk that
default parameters from Factory differ from those of the Corrector
classes. Decision: avoid default critical parameter in LabelledChemical
and in *Corrector classes.

* removed several default values in LabelledChemical and *Corrector
* consequently, had to explicitly ask for:
   * "orbitrap" resolution_formula_code
   * no derivative_formula
   * perfect tracer purity
   * default data_isotopes